+++ CHANGELOG.rst
@@ -20,6 +20,10 @@ Added
 - ``FromConfigMixin.from_config`` with ``config_read_mode_fsspec_enabled=True``
   now supports fsspec config loading including the resolving of relative paths
   (`#918 <https://github.com/omni-us/jsonargparse/pull/918>`__).
+- Add :obj:`.Unset` sentinel and ``unset_sentinel`` setting in
+  ``.set_parsing_settings`` to distinguish between arguments not provided and
+  those explicitly set to ``null`` (`#909
+  <https://github.com/omni-us/jsonargparse/pull/909>`__).
 
 Fixed
 ^^^^^
@@ -31,6 +35,16 @@ Changed
 ^^^^^^^
 - Drop support for Python 3.9. The minimum supported Python version is now
   3.10 (`#916 <https://github.com/omni-us/jsonargparse/pull/916>`__).
+
+Deprecated
+^^^^^^^^^^
+- ``skip_none`` parameter of :meth:`dump <.ArgumentParser.dump>`, :meth:`save
+  <.ArgumentParser.save>`, and :meth:`validate <.ArgumentParser.validate>` was
+  deprecated and will be removed in v5.0.0. Use ``skip_unset`` instead (`#909
+  <https://github.com/omni-us/jsonargparse/pull/909>`__).
+- ``skip_null`` flag for ``--print_config`` was deprecated and will be removed
+  in v5.0.0. Use ``skip_unset`` instead (`#909
+  <https://github.com/omni-us/jsonargparse/pull/909>`__).
 
 
 v4.49.0 (2026-05-15)